+++ src/constants.ts
@@ -1,0 +1,5 @@
+export const GUARD_TYPE = {
+  guard: 'guard',
+  and: 'and',
+  or: 'or',
+} as const;